# [baremetal] AES remove forward reverse tables to save ROM