… library/version_features.c
… scripts/config.pl