## Description

Replace AES forward/reverse tables with instant calculation to save ROM.

__Impact on code-size:__

| | GCC | ARMC5 | ARMC6 |
| --- | --- | --- | --- |
| `libmbedcrypto.a` before | 20915 | 21387 | 23477 |
| `libmbedcrypto.a` after | 18875 | 19207 | 21441 |
| gain Total | 2040 | 2180 | 2036 |

## Status
**READY**

## Additional comments
This PR contains commits from PR #2888