@mpg 2019-10-18
@artokin Did you make any measurement of the effect on speed, preferably on a target of interest? (We have a crypto brenchmark program for Mbed OS which could be useful: https://github.com/ARMmbed/mbed-os-example-tls/tree/master/benchmark - or you could roll your own.) The main benefit of larger tables is speed, so it would be interesting to know how much we loose by removing them.