@artokin 2019-10-21
@mpg , bench-marking results attached, only AES test results copied here as there were no significant difference in the other tests:

__Benchmarking with ARMC5:__

| test | before | after |
| --- | --- | --- |
| AES-CBC-128      | 536 KB/s | 160 KB/s
| AES-CBC-192      | 472 KB/s | 134 KB/s
| AES-CBC-256      | 422 KB/s | 115 KB/s
| AES-CTR-128      | 564 KB/s | 162 KB/s
| AES-CTR-192      | 494 KB/s | 136 KB/s
| AES-CTR-256      | 437 KB/s | 116 KB/s
| AES-GCM-128      | 317 KB/s | 132 KB/s
| AES-GCM-192      | 293 KB/s | 114 KB/s
| AES-GCM-256      | 272 KB/s | 100 KB/s
| AES-CCM-128      | 240 KB/s | 77 KB/s
| AES-CCM-192      | 214 KB/s | 65 KB/s
| AES-CCM-256      | 192 KB/s | 56 KB/s
| AES-CMAC-128     | 474 KB/s | 152 KB/s
| AES-CMAC-192     | 422 KB/s | 128 KB/s
| AES-CMAC-256     | 379 KB/s | 111 KB/s
| AES-CMAC-PRF-128 | 473 KB/s | 152 KB/s