@mpg 2019-10-22
Thanks for making the measurements!

Wow, that's a pretty big performance hit! Looking only at the line for AES-CCM-128, which I think is the only one we care about, the tableless version is 3 times slower.

This makes me doubt if this codesize optimisation is really worth it. Currently it's measured at about 2kB, but #2890 already achieved a gain of 2.5kB without any performance hit, and I doubt the gains between those two PRs are additive (ie, this PR's gain would probably be smaller if measured on top of #2890).

@sbutcher-arm What do you think? Is it acceptable to make AES-CCM-128 3 times slower for a gain of 2kB of code size (or maybe less)? Or should we instead focus on other ways to save code size, that don't impact performance as much?