@artokin 2019-10-22
This was rebased due merge conflict in `scripts/config.pl`.
Please note, a new rebase will be needed when once #2890 and #2891 are merged.