@mpg 2019-10-22
@artokin Out of curiosity, could you try to do a quick local rebase over 2891 (or merge 2891 with this one in a local branch) and measure what the codesize improvement is when enabling `MBEDTLS_AES_SBOX_ONLY` on top of `MBEDTLS_AES_ONLY_ENCRYPT`? I suspect it's likely to be about half of the currently measured one, so about 1kB, but it's easy to guess wrong with these things, so an actual measurement would be better than my guess :)