@sbutcher-arm 2019-10-22
I can see your concerns @mpg, but to make a decision I think we need absolute figures.

@artokin - Once we've merged the *other* AES PR's, can you measure again the improvement in codesize and cost in RAM? Once we have that, we can make a decision.