@artokin 2019-10-22
I made a merge in my workarea and here are the initial ROM figures:

__Configuration impact to code size using ARMC5_5.06 update 6__

| Configuration | ROM usage | Gain to baremetal |
|--------|-----------|------|
| baremetal | 42762 | 0 |
| AES 128-bit key | 42512  | -250  |
| AES 128-bit key + encrypt_only | 39894 | -2868 |
| AES 128-bit key + encrypt_only + SBOX Table| 38728 | -4034 |