@mpg 2019-10-23
@artokin Thanks for those doing those measurements! So if I'm reading the table correctly, the additional gain of table-less AES on top of encrypt-only is 1166 bytes.

To summarize, here's the trade-off that this PR implements:

**Benefit:** saving 1166 bytes of flash.
**Costs:**

* Dividing performance by 3 for AES-CCM-128 (from 240 to 77kB/s on the target)
* Consume 256 more bytes of static RAM (fixed cost for the application, not per context or operation, but permanent (not on stack or heap)).

@artokin Do you think this summary is accurate?

@sbutcher-arm Considering that summary, do you think we should pursue this or park it?

My opinion is that the performance cost is so high that we should give priority to other ways of saving code size, that don't impact performance as much - we still have ideas for such optimisations.