@artokin 2019-10-23
@mpg , your summary is excellent!